+++ CHANGELOG.md
@@ -4,9 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2019-05-23
+### Changed
+- `MicroService-Call` constructs, don't need parametres anymore
+
 ## [Unreleased] - 2019-05-22
 ### Added
 - Project inited
-- MicroService-Call added
-- Divided MicroService-Call and Router-Fetcher
-- Tests
+- `MicroService-Call` module added
+- `Unit Tests` added
+### Removed
+- `Router-Fetcher`moved to an independent package.